fetch data using react-query
